Issue 1340: studentFeedbackSessionResults: view results for MCQ question
Update Issue 1340
Status: Delivered
For now I only display the choice chosen by the giver as text.

Is it better if students can see a compiled result (e.g. how many people choose a certain option) or is it fine as it is?

I assume the instructor's view should have a way to get the compiled result?

--HG--
branch : mcq


Former-commit-id: e61416caf064b482df97158e04230d982769caec
Former-commit-id: 27d33e7c6c17a7c2a9a5fea088dd3c7be33709fe
Former-commit-id: 88c0111f1f7de5adf4da4297facffc8494154a42