Fix poor color of 'more info' link beside contrib question
From what I see, all other question templates use the
`FEEDBACK_QUESTION_ADDITIONAL_INFO` for the "more info"
link beside the question title. So this commit removes the
class for this link that gave it light yellow color
(link-in-dark-bg). Instead, create a new fragment that
the contribution question can use for the "more info"
link in the panel and have that link use "link-in-dark-bg" instead.