Fix response maps to account for when responses are not grouped by team
When the responses are not sorted by the primary participant's team,
the grouping done by the various response maps may be incorrect as some
depended on iterating through the participants in team order. Therefore,
the initialisation of the data structures have been changed to not rely on the sorted order of the responses.

However, the order of responses still determines the order of the
secondary participant panels and response panels on the results page.