[5261] Group anonymous responses together in results page
* Modify comparators to treat anonymous names separately from visible ones

* Fix response maps to account for when responses are not grouped by team

When the responses are not sorted by the primary participant's team,
the grouping done by the various response maps may be incorrect as some
depended on iterating through the participants in team order. Therefore,
the initialisation of the data structures have been changed to not rely on the sorted order of the responses.

However, the order of responses still determines the order of the
secondary participant panels and response panels on the results page.

* Update html comparisons for student results

* Fix non-participants panels when not group by team

When not grouped by team, construct missing participants' panels when there is a change in section instead of team. This is because we can no longer rely on the responses getting grouped by team names

* Update tests for instructor results

* Fix failing tests

caused by anonymous responses getting reordered for students results page, and recipient indexes (of the response index) getting swapped for RGQ and RQG.

* Fix spacing