Changing default collapse section button visibility in sectionPanel.tag and debug for InstructorFeedbackResultsPageUiTest.java
For collapse section button, the original button is visible by default as
Ajax auto loads all the panels. However, now as we do not auto load all
panels, the button should be invisible by default.

For debugging InstructorFeedbackResultsPageUiTest.java, as now the
panels do not auto load, in all the test the expand panels button is
clicked(is button is not present, clicking all panels heading). Run god
mode to change the corresponding html page running with Firefox 46.0.1
on Mac OS 10.11.6