Change rtf toolbar's inline property to false in instructorFeedbackEdit.es6:
Change the toolbar’s inline property to false in enableEditFS(),
enableQuestion(questionNum) and enableNewQuestion() so that toolbar is
fixed when question/fs details are being edited or new question is
added.

Keep the toolbar’s inline property as true for disableEditFS(),
disableQuestion(questionNum) and $(document).ready() so that toolbar
remains hidden when fields are not being edited.