Moved extractFeedbackSessionData() to InstructorFeedbacksPageAction.java
extractFeedbackSessionData() was redundant in InstructorFeedbackEditSaveAction.java and InstructorFeedbackAddAction.java - moved to parent class and added a check for if the method is creating or not, so it does not overwrite the date value when editing.

InstructorFeedbackEditSaveAction.java was inheriting directly from the Action class when it should inherit from InstructorFeedbackPageAction.java - since it's a InstructorFeedbackPage action.

InstructorFeedbackAddAction.java - just removed the extractFeedbackSessionData().