Add IntelliJ specific build folder
IntelliJ possibly still consider Gradle as a "single source of truth".
https://intellij-support.jetbrains.com/hc/en-us/community/posts/206810455-JetGradle-tool-window-not-available?page=1#community_comment_206256495

However, due to leaky abstractions such as differing exploded war directory
used in Gradle and IntelliJ and temporary folders, IntelliJ is unable to use
Gradle as a "single source of truth" properly, thus builds from Gradle and
IntelliJ may not match so it is better to separate the folders.