Add tasks to make IntelliJ build path match Gradle
IntelliJ creates possibly the wrong settings for compile output path in the
root module. This might be a bug or perhaps intended behaviour. In any case,
the settings causes resources to be built to two different folder thus
resulting in duplicates.

This commit thus make the settings match Gradle build output. However, it is
done in an unsafe way.

IntelliJ does not have the concept of saving and after every Gradle sync it
attempts to change the root module settings (to the possibly wrong one).
And a possible way to override IntelliJ's settings is to create a conflict with
a line IntelliJ modified, delete the file and regenerate it, but this can fail
in all kinds of ways because it is not documented behavior and we are basically
fighting the cache. The solution also have not been extensively tested.

Another problem is that IntelliJ auto creates an exploded war directory if the
war plugin in build.gradle has been detected. However, the `war` plugin does
not create an exploded war directory. The `appengine` plugin does however, but
the path is not the same as the one IntelliJ creates. One could say this is the
problem of leaky abstractions as IntelliJ might need an exploded war directory
but Gradle does not have them by default, so IntelliJ creates one for itself
and appengine create another for itself. Thus this commit also changes the
exploded war output directory to match the `appengine` plugin.