Change instructions for setup of run configuration
Change the instructions for setting up a `Google AppEngine Dev Server` in
IntelliJ as the `assemble` Gradle task is no longer required for building.

Shortcut for stopping the dev server is also removed to be consistent with all
other instructions. Instructions only show how to do things in the long form
and users can easily discover the shortcuts on their own. Furthermore, the
shortcut is also `Windows` and `Keymap` specific.