Standardize project name to allow different project root
A different project root directory results in a different project name.
However, the project name should always be the same. Gradle does not support
specifying the project name in the build script so we specify them in
`settings.gradle`.

Additionally, we also set the same project name in IntelliJ IDEA for
consistency.

It is to be noted that that IntelliJ IDEA have the same behaviour when creating
projects with project names different from project root directory.