Drop support for ES7 and above
ES7 and ES8 are not supported even in the latest of Eclipse.

While that should not be the decision-making point, another is that there is no new feature (see http://exploringjs.com/es2016-es2017/) from the new specs that we currently use, other than the trailing comma in function calls which is easily dispensable.