Make selecting new questions wait to be ready
We should always wait for an object to be ready before doing
something else.

In this case, one issue is that the new question panel will be
scrolled to when creating new questions, and if in the meantime we
try to scroll to another element they can interfere with each other.