Improve reliability of some tests related to status messages.
Many status messages are unexpectedly empty in tests. This is because
status messages is not unique to a single request, but for an entire
session. However, they work like they are request-unique.

The implementation is leaky because the status message for a request
can be cleared by another request with the right timing. For more
details on how status messages work, see:
https://github.com/TEAMMATES/teammates/blob/8ad5768d4007a1fd5e8eca729975d779840b1126/src/main/java/teammates/ui/controller/FeedbackSubmissionEditPageAction.java#L89-L106

The tests that are to be fixed because of the reason above and the steps
to replicate the problem is as follows:
1. Add or edit a question.
2. Select a visibility message option or custom feedback path option.
2. An AJAX request will be sent to retrieve the new visibility message.
3. Before the AJAX request completes, add or save the question.
4. If the timing is right, the AJAX request will clear the status message
for the added or saved question.

The tracking issue for this problem can be found in the PR description.

In order to fix the tests, the following approach is taken.
1. Add or edit a question.
2. Select a visibility message option or custom feedback path option.
2. An AJAX request will be sent to retrieve the new visibility message.
3. The test would wait for the AJAX request to complete before adding
   or saving the question.
4. The AJAX request will then never clear the status message for
   the added or saved question.