Fix `fillTextBox` and `fillRichTextEditor`
Fix `fillTextBox` by firing `change` event when text input changes if Firefox not in focus
because of a Firefox bug.

Fix broken insertion of data in `fillRichTextEditor`. In particular,
the removal of redundant tabs when inserting data in TinyMCE.

The additional tabs required the modification of a test in order for it to pass.
We end up masking the actual bug where the cause is that the backslash is inserted as
a Javascript String, which causes '\&' to turn into '&' because the backslash then
becomes an escape modifier which effectively escapes nothing.

Unlike `fillTextBox`, `fillRichTextEditor` is not affected by focus problems in Firefox
as some events are specific to TinyMCE, for example the TinyMCE `Change` event is fired
immediately when we set a new value.

We can then remove code that fires `change` events manually.

Future work can be to reduce the disparity in code for filling other editable content as compared to TinyMCE
content.

For instance the following code explores how to improve `clearAndSetNewValueForTinyMce`:
WebElement tinyMceEditableArea = (WebElement) executeScript("return tinyMCE.get('" + id + "').getBody();");
switchToTinyMceIframeIfNotInline(id);
tinyMceEditableArea.sendKeys(Keys.TAB);

// Make sure we switch back to parent frame if required so other Selenium calls will be operating in
// the correct frame.
browser.driver.switchTo().parentFrame();

private void switchToTinyMceIframeIfNotInline(String tinyMceId) {
    @SuppressWarnings("unchecked")
    Map<String, Object> tinyMceInstanceData = (Map<String, Object>) executeScript(
            "const tinyMceEditorInstance = tinyMCE.get(arguments[0]);"
                    + "return  {"
                    + "       inline: tinyMceEditorInstance.inline === true,"
                    + "       iframe: tinyMceEditorInstance.iframeElement"
                    + "}", tinyMceId);

    if (!((Boolean) tinyMceInstanceData.get("inline"))) {
        final WebElement iframeElement = (WebElement) tinyMceInstanceData.get("iframe");
        browser.driver.switchTo().frame(iframeElement);
    }
}