Edit "Create Session" section
Added section that gives overview of how to set up a feedback session

Made the section a single column instead of using a 2-column layout.
The 2-column layout was a bit wordy and distracting. To be fair, this
layout is also quite wordy, but that's due to the entire page being
one column -- a possible issue for the future.

Moved optional advanced information into the same list as required
"quick start" information -- additional information is still part of
the same workflow as required information. I included an "(Optional)"
line so readers who are uninterested can skim or skip this bullet point.

Removed use of grayed out words. They impeded readability, and were
at the same time quite distracting due to the format of the list.
Am open to putting them back in, possibly in a different format.
Alternatively, we could bold the keywords (e.g. "grace period"), but
I think that is also quite distracting.