Edit "Adding Questions" section
Changed section title to "How do I add questions to a session?". All section
headings will be phrased as questions.

Removed links to each question type from this section, providing only a
link to the section on question types. The rationale for this is that as
a guide to adding questions, this section should provide in the most concise
terms possible a guide to adding questions. Users who are interested to
find out more about the question types can follow the link to learn more.

Removed 2 example UI elements. The first element showed a preview of the
question from the perspective of the question answerer, which does not
answer the question "How do I add questions to my session?"

The second UI element was quite similar to the first element, showing
different choices of visibility/feedback path hints. Since these are
available from the dropdown, it suffices to indicate where the
dropdown is, and allow users to explore the options on their own.