Edit "Results" section
1) Change section header into a question
2) Removed a line about filtering results using the "filter function" as
the current UI does not contain a filter function.
3) Unbolded sections of explanatory text. These parts are not the main
point of the section, and are also quite verbose, so bolding them all
draws attention away from instructions.
3) Todo: Check if the existing UI element for "Edit View" can be fixed
4) Todo: Set up UI Element minimizer; the UI elements in this section
are very large. This impedes readability, as the next block of text is
very far away and the element is quite complicated. Readers may get lost
trying to understand what the UI element is showing them. We can consider
not including these elements, but I think they are useful. One trade-off
is to hide the elements under a dropdown, so that they will only appear
if the reader is interested in seeing them.