Move all questions into collapsibles
The aim of this move is to remove the table of contents entirely, and
have users scroll through and click to uncollapse the relevant questions.
This allows them to keep track of 'where' they are in the table of contents.

Errors: some UI elements (e.g. preview as panel) do not resize well. We
can manually resize these elements, but we should also think about a more
long-term solution. Suggestions welcome!

Also, not sure how readable the nested collapsibles are.