Move "Sections" section into "Courses" section
Talking about sections doesn't really make sense
without the context that sections are features of courses.
Moving the "sections" questions into "courses" may help
users to find the relevant questions more quickly as
opposed to when information was segregated by the technical
function of the feature.

Also renamed InstructorHelpAddEditInstructors to
InstructorHelpCourses, as it is now used for more general
questions.