Change int members to long
This avoids awkward casting from int to long in the methods that have
been updated to use streams.

totalOngoingSessions will be of type long, but be assigned an int type
variable. Since the range of int is a strict subset of long, Java does
not complain about the implicit cast.