Add ability to create links that only toggle collapsibles open
If a link in the instructorHelp file contains the class 'collapse-link'
and a data-target attribute, then the collapsible identified by
data-target will be shown. If the link is clicked multiple times,
the collapsible will not be hidden; i.e. the toggle only shows, but
does not hide, the content in the collapsible.

This class is for the use of links in the instructorHelp page
which link to collapsibles that might already be open. If the
collapsible is already open, we don't want to toggle it closed,
which the current data- attributes do.x