Edit "Comments" section
Section content edited for clarity, and put inside collapsibles. The
question on searching comments was deleted as it is a duplicate of the
question in the "search" section.

During reorganization, the "search" section will be located in the "sessions"
subheader, so will not be necessary to duplicate data on searching for
comments.