Update GettingStarted
Added more granular information about what can be done in TEAMMATES,
for instance adding comments to responses. This is done to consolidate
all information into a concise and well-formed introduction to TEAMMATES.

Split section about "Learning more" into two separate sections. This is
to show new instructors that there are added functions to TEAMMATES
outside of conducting and managing feedback sessions. In addition, the
information on creating and managing sessions is separated to show two
distinct steps.

Added links to instructorHelp sessions as far as possible. This is done
to allow "sequential access" users to access information as they go
through the "Getting Started" page. Ideally, this information would open
up in a different page; can consider making "Getting Started" a separate
page as well.