Add sectionDetail to all the functions called in FeedbackResponsesLogic, FeedbackSessionsLogic and FeedbackResponsesDb.
Currently, all functions with section as a parameter will have a sectionDetail parameter to add to it. However, this can be refactored in a later stage as section is null sometimes and sectionDetail should be null in these cases. There are also cases where sectionDetail does not apply to the section, hence it can be refactored to be null.