Remove getFeedbackSessionResultsForInstructorFromSection and getFeedbackSessionResultsForInstructorToSection which are not called anywhere.
I suspect these functions may be used very long ago to filter between resultsToSection and resultsFromSection. However now that there is a filter option to choose which of the option is selected through the sectionDetail option. InstructorFeedbackResultsPageAction only calls getFeedbackSessionResultsForInstructorInSection which will initialize the params accordingly to the selectedSectionDetail. Hence the function toSection and fromSection can be safely removed.