Instructor help page: Fix broken links
The current behaviour of the <a> elements allows user's to open in a
new tab, acting as links.

Let's
* convert the role of the <a> elements to role = "button"
* style the elements such that it still looks the same as a link

The bug lies in [tmRouterLink] = "" which converts the <a> to a link,
additionally it only links back to the home page. To remain consistent
with the current styling, the button tag is styled using .scss

Refer to the use of button classes to trigger in-page-functionality
https://getbootstrap.com/docs/4.0/components/buttons/#button-tags