Create component for a row of new instructor data
There is currently no (easy) way to maintain some sort of state or any variables for each row of new instructor data.

An edit new instructor data feature would require each row to maintain a state of whether it is being edited, how it was edited, and where it was edited. However, the current implementation does not (easily) allow this.

Let's create a component for a row of new instructor data. This creates a class for a row of new instructor data. This can be instantiated, allowing each one to maintain a state and some variables.