Remove course ID set
Currently, there is a set that tracks all the course IDs. During
migration, we will need to track course IDs for both courses that
are migrated and courses that have not been migrated yet.

If we keep the course ID set, then during migration, there will be
two course ID sets, as well as two course ID feedback session maps.
This may be messy.

The course IDs can be tracked in the course ID feedback session
maps.

Let's remove the course ID set so that it does not need to be
duplicated. This may make the code less messy during migration.