[#12048] Migrate GetOngoingSessionsAction for V9 (#12710)
* Remove exception message magic strings

* Extract duplicate parse code as method

* Extract time parameter validation method

* Remove course ID set

Currently, there is a set that tracks all the course IDs. During
migration, we will need to track course IDs for both courses that
are migrated and courses that have not been migrated yet.

If we keep the course ID set, then during migration, there will be
two course ID sets, as well as two course ID feedback session maps.
This may be messy.

The course IDs can be tracked in the course ID feedback session
maps.

Let's remove the course ID set so that it does not need to be
duplicated. This may make the code less messy during migration.

* Extract output creation as method

* Extract course ID to feedback sessions map creation method

* Extract institute to feedback sessions map creation method

* Remove dependency on account in ongoing session

* Add method to get ongoing feedback sessions

* Add method to get ongoing sessions in feedback sessions logic

* Add method to get ongoing sessions in the logic class

* Inject mock datastore logic class

* Add method to check whether a feedback session is waiting to open

* Support SQL feedback session entity in OngoingSession output

* Update action to get ongoing sessions for migration to SQL

* Fix total ongoing sessions counting

* Fix typo errors in typical case unit test

* Improve clarity of unit tests

* Test when course is migrated but not the instructor account

* Remove unnecessary comments